Moved global code to utils
